Rework peripherals components system
